Refactor the interval-api to make it more organized (#16462)
[AB#5031](https://dev.azure.com/fluidframework/internal/_workitems/edit/5031)

Please refer to the description in
https://github.com/microsoft/FluidFramework/pull/16138, the content is
exact the same except for some breaking changes (which will be committed
to next). It includes a changeset as well.

---------
