+++ packages/dds/sequence/src/revertibles.ts
@@ -20,7 +20,7 @@ import {
 	SortedSet,
 	getSlideToSegoff,
 } from "@fluidframework/merge-tree";
-import { IntervalOpType, SequenceInterval } from "./intervalCollection";
+import { IntervalOpType, SequenceInterval } from "./intervals";
 import { SharedString, SharedStringSegment } from "./sharedString";
 import { ISequenceDeltaRange, SequenceDeltaEvent } from "./sequenceDeltaEvent";
