+++ packages/dds/sequence/src/test/intervalCollection.spec.ts
@@ -21,14 +21,13 @@ import {
 import { LoggingError } from "@fluidframework/telemetry-utils";
 import { SharedString } from "../sharedString";
 import { SharedStringFactory } from "../sequenceFactory";
+import { IIntervalCollection, IntervalIndex } from "../intervalCollection";
 import {
-	IIntervalCollection,
 	IntervalStickiness,
 	IntervalType,
 	SequenceInterval,
-	IntervalIndex,
 	ISerializableInterval,
-} from "../intervalCollection";
+} from "../intervals";
 
 class MockIntervalIndex<TInterval extends ISerializableInterval>
 	implements IntervalIndex<TInterval>
