+++ packages/dds/sequence/src/test/intervalRebasing.spec.ts
@@ -9,7 +9,7 @@ import {
 	MockFluidDataStoreRuntime,
 	MockStorage,
 } from "@fluidframework/test-runtime-utils";
-import { IntervalType } from "../intervalCollection";
+import { IntervalType } from "../intervals";
 import { SharedStringFactory } from "../sequenceFactory";
 import { SharedString } from "../sharedString";
 import { assertConsistent, Client } from "./intervalUtils";
