+++ packages/dds/sequence/src/test/startpointInRangeIndex.spec.ts
@@ -6,12 +6,8 @@
 import { strict as assert } from "assert";
 import { makeRandom } from "@fluid-internal/stochastic-test-utils";
 import { Client } from "@fluidframework/merge-tree";
-import {
-	Interval,
-	IStartpointInRangeIndex,
-	createStartpointInRangeIndex,
-	intervalHelpers,
-} from "../intervalCollection";
+import { IStartpointInRangeIndex, createStartpointInRangeIndex } from "../intervalCollection";
+import { Interval, intervalHelpers } from "../intervals";
 import {
 	assertPlainNumberIntervals,
 	createTestInterval,
